requirements.txt
psycopg2-binary
python-dotenv
fhir.resources
requests